+++ sasl.go
@@ -21,11 +21,14 @@ const (
 	frameTypeSASL = 0x1
 )
 
+// SASLType represents a SASL configuration to use during authentication.
+type SASLType func(c *conn) error
+
 // ConnSASLPlain enables SASL PLAIN authentication for the connection.
 //
 // SASL PLAIN transmits credentials in plain text and should only be used
 // on TLS/SSL enabled connection.
-func ConnSASLPlain(username, password string) ConnOption {
+func SASLTypePlain(username, password string) SASLType {
 	// TODO: how widely used is hostname? should it be supported
 	return func(c *conn) error {
 		// make handlers map if no other mechanism has
@@ -58,7 +61,7 @@ func ConnSASLPlain(username, password string) ConnOption {
 }
 
 // ConnSASLAnonymous enables SASL ANONYMOUS authentication for the connection.
-func ConnSASLAnonymous() ConnOption {
+func SASLTypeAnonymous() SASLType {
 	return func(c *conn) error {
 		// make handlers map if no other mechanism has
 		if c.saslHandlers == nil {
@@ -90,7 +93,7 @@ func ConnSASLAnonymous() ConnOption {
 // ConnSASLExternal enables SASL EXTERNAL authentication for the connection.
 // The value for resp is dependent on the type of authentication (empty string is common for TLS).
 // See https://datatracker.ietf.org/doc/html/rfc4422#appendix-A for additional info.
-func ConnSASLExternal(resp string) ConnOption {
+func SASLTypeExternal(resp string) SASLType {
 	return func(c *conn) error {
 		// make handlers map if no other mechanism has
 		if c.saslHandlers == nil {
@@ -129,7 +132,7 @@ func ConnSASLExternal(resp string) ConnOption {
 //
 // SASL XOAUTH2 transmits the bearer in plain text and should only be used
 // on TLS/SSL enabled connection.
-func ConnSASLXOAUTH2(username, bearer string, saslMaxFrameSizeOverride uint32) ConnOption {
+func SASLTypeXOAUTH2(username, bearer string, saslMaxFrameSizeOverride uint32) SASLType {
 	return func(c *conn) error {
 		// make handlers map if no other mechanism has
 		if c.saslHandlers == nil {