+++ sasl_test.go
@@ -103,9 +103,10 @@ func TestConnSASLXOAUTH2AuthSuccess(t *testing.T) {
 	}
 
 	c := testconn.New(buf)
-	client, err := New(c,
-		ConnSASLXOAUTH2("someuser@example.com", "ya29.vF9dft4qmTc2Nvb3RlckBhdHRhdmlzdGEuY29tCg", 512),
-		ConnIdleTimeout(10*time.Minute))
+	client, err := New(c, &ConnOptions{
+		IdleTimeout: 10 * time.Minute,
+		SASLType:    SASLTypeXOAUTH2("someuser@example.com", "ya29.vF9dft4qmTc2Nvb3RlckBhdHRhdmlzdGEuY29tCg", 512),
+	})
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -132,9 +133,10 @@ func TestConnSASLXOAUTH2AuthFail(t *testing.T) {
 	}
 
 	c := testconn.New(buf)
-	client, err := New(c,
-		ConnSASLXOAUTH2("someuser@example.com", "ya29.vF9dft4qmTc2Nvb3RlckBhdHRhdmlzdGEuY29tCg", 512),
-		ConnIdleTimeout(10*time.Minute))
+	client, err := New(c, &ConnOptions{
+		IdleTimeout: 10 * time.Minute,
+		SASLType:    SASLTypeXOAUTH2("someuser@example.com", "ya29.vF9dft4qmTc2Nvb3RlckBhdHRhdmlzdGEuY29tCg", 512),
+	})
 	if err != nil {
 		defer client.Close()
 	}
@@ -171,9 +173,10 @@ func TestConnSASLXOAUTH2AuthFailWithErrorResponse(t *testing.T) {
 	}
 
 	c := testconn.New(buf)
-	client, err := New(c,
-		ConnSASLXOAUTH2("someuser@example.com", "ya29.vF9dft4qmTc2Nvb3RlckBhdHRhdmlzdGEuY29tCg", 512),
-		ConnIdleTimeout(10*time.Minute))
+	client, err := New(c, &ConnOptions{
+		IdleTimeout: 10 * time.Minute,
+		SASLType:    SASLTypeXOAUTH2("someuser@example.com", "ya29.vF9dft4qmTc2Nvb3RlckBhdHRhdmlzdGEuY29tCg", 512),
+	})
 	if err != nil {
 		defer client.Close()
 	}
@@ -210,9 +213,10 @@ func TestConnSASLXOAUTH2AuthFailsAdditionalErrorResponse(t *testing.T) {
 	}
 
 	c := testconn.New(buf)
-	client, err := New(c,
-		ConnSASLXOAUTH2("someuser@example.com", "ya29.vF9dft4qmTc2Nvb3RlckBhdHRhdmlzdGEuY29tCg", 512),
-		ConnIdleTimeout(10*time.Minute))
+	client, err := New(c, &ConnOptions{
+		IdleTimeout: 10 * time.Minute,
+		SASLType:    SASLTypeXOAUTH2("someuser@example.com", "ya29.vF9dft4qmTc2Nvb3RlckBhdHRhdmlzdGEuY29tCg", 512),
+	})
 	if err != nil {
 		defer client.Close()
 	}
@@ -250,9 +254,10 @@ func TestConnSASLExternal(t *testing.T) {
 	}
 
 	c := testconn.New(buf)
-	client, err := New(c,
-		ConnSASLExternal(""),
-		ConnIdleTimeout(10*time.Minute))
+	client, err := New(c, &ConnOptions{
+		IdleTimeout: 10 * time.Minute,
+		SASLType:    SASLTypeExternal(""),
+	})
 	if err != nil {
 		t.Fatal(err)
 	}