+++ sender_test.go
@@ -17,7 +17,7 @@ import (
 func TestSenderInvalidOptions(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(ModeUnsettled))
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -67,7 +67,7 @@ func TestSenderMethodsNoSend(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -104,7 +104,7 @@ func TestSenderMethodsNoSend(t *testing.T) {
 func TestSenderSendOnClosed(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(ModeUnsettled))
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -130,7 +130,7 @@ func TestSenderSendOnClosed(t *testing.T) {
 func TestSenderSendOnSessionClosed(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(ModeUnsettled))
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -156,7 +156,7 @@ func TestSenderSendOnSessionClosed(t *testing.T) {
 func TestSenderSendOnConnClosed(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(ModeUnsettled))
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -182,7 +182,7 @@ func TestSenderSendOnConnClosed(t *testing.T) {
 func TestSenderSendOnDetached(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(ModeUnsettled))
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -238,7 +238,7 @@ func TestSenderAttachError(t *testing.T) {
 		}
 	}
 	netConn := mocks.NewNetConn(responder)
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -286,7 +286,7 @@ func TestSenderAttachError(t *testing.T) {
 func TestSenderSendMismatchedModes(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(ModeUnsettled))
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -329,7 +329,7 @@ func TestSenderSendSuccess(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -374,7 +374,7 @@ func TestSenderSendSettled(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -415,7 +415,7 @@ func TestSenderSendRejected(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -480,7 +480,7 @@ func TestSenderSendRejectedNoDetach(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -529,7 +529,7 @@ func TestSenderSendDetached(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -558,7 +558,7 @@ func TestSenderSendDetached(t *testing.T) {
 func TestSenderSendTimeout(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(ModeUnsettled))
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -613,7 +613,7 @@ func TestSenderSendMsgTooBig(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -649,7 +649,7 @@ func TestSenderSendTagTooBig(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -721,7 +721,7 @@ func TestSenderSendMultiTransfer(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -752,7 +752,7 @@ func TestSenderSendMultiTransfer(t *testing.T) {
 func TestSenderConnReaderError(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(ModeUnsettled))
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -785,7 +785,7 @@ func TestSenderConnReaderError(t *testing.T) {
 func TestSenderConnWriterError(t *testing.T) {
 	netConn := mocks.NewNetConn(senderFrameHandlerNoUnhandled(ModeUnsettled))
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
@@ -841,7 +841,7 @@ func TestSenderFlowFrameWithEcho(t *testing.T) {
 	}
 	netConn := mocks.NewNetConn(responder)
 
-	client, err := New(netConn)
+	client, err := New(netConn, nil)
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)