+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
 * Added `ConnectionError` type that's returned when a connection is no longer functional.
 * Added `LinkTargetCapabilities()` option to specify desired target capabilities.
+* Added `SASLType` used when configuring the SASL authentication mechanism.
 
 ### Breaking Changes
 * Removed `ErrConnClosed` and `ErrTimeout` sentinel error types.
@@ -15,6 +16,11 @@
   * `AMQPAddress`, `AMQPMessageID`, `AMQPSymbol`, `AMQPSequenceNumber`, `AMQPBinary`
 * Various `Default*` constants are no longer exported.
 * The args to `Receiver.ModifyMessage()` have changed.
+* The "variadic config" pattern for `Client` constructors has been replaced with a struct-based config.
+  * This removes the `ConnOption` type and all of the associated configuration funcs.
+  * The `ConnTLS()` option was removed as part of this change.
+* The `Dial()` and `New()` constructors now require an `*ConnOptions` parameter.
+* The various SASL configuration funcs have been slightly renamed.
 
 ### Bugs Fixed
 * Fixed potential panic in `muxHandleFrame()` when checking for manual creditor.